sistemaEscolar/src/main/webapp/app/entities/plano-ensino/delete/plano-ensino-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, fakeAsync, inject, tick } from '@angular/core/testing';
import { HttpResponse, provideHttpClient } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { PlanoEnsinoService } from '../service/plano-ensino.service';

import { PlanoEnsinoDeleteDialogComponent } from './plano-ensino-delete-dialog.component';

describe('PlanoEnsino Management Delete Component', () => {
  let comp: PlanoEnsinoDeleteDialogComponent;
  let fixture: ComponentFixture<PlanoEnsinoDeleteDialogComponent>;
  let service: PlanoEnsinoService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [PlanoEnsinoDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(PlanoEnsinoDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(PlanoEnsinoDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(PlanoEnsinoService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
